refactor: cihaz detayı sekme menüsünü okunaklı hale getir ve ortak komponente çıkar
Cihaz Bilgileri sayfasındaki sekme menüsü koyu lacivert arka plan üzerine
okunaksız (görünmeyen) metinle render oluyordu. Personel detayındaki okunaklı
pill-tarzı sekme menüsü (Kişisel Bilgiler / Çalışma Bilgileri / ...) referans
alınarak yeni bir DetailPageTabs komponenti çıkarıldı ve hem Cihaz hem Personel
detay sayfaları bu ortak komponenti kullanacak şekilde güncellendi.

Codebase'de bu eski/okunaksız sekme deseni sadece bu iki dosyada vardı; başka
sayfa etkilenmedi.
