+++ resources/js/Pages/Modules/HumanResources/Employee/ShowPage.vue
@@ -5,8 +5,8 @@ import {Link, router } from "@inertiajs/vue3";
 // Components
 import SimpleButton from "@/Components/Button/SimpleButton.vue"
 import Avatar from "@/Components/Avatar/Avatar.vue";
-import SelectInput from "@/Components/Form/SelectInput.vue";
 import HelpButton from "@/Components/Help/HelpButton.vue";
+import DetailPageTabs from "@/Components/Navigation/DetailPageTabs.vue";
 
 // Props
 const props = defineProps({
@@ -15,14 +15,9 @@ const props = defineProps({
 
 // Multi-lang
 import Translates from "./translates"
-import {ref} from "vue";
 
 const {t, tm} = Translates();
 
-const selectedTab = ref(route().current());
-const changePage = ()=>{
-  router.visit(route(selectedTab.value, props.employee.id));
-}
 /*Tabs*/
 const tabs = [
   {
@@ -102,25 +97,7 @@ const handleDelete = () => {
     </template>
 
     <div>
-      <!--Tabs-->
-      <div class="hidden md:flex select-none pb-3 -mb-3 scrollbar-thin scrollbar-thumb-rose-500">
-        <template v-for="i in tabs">
-          <Link :href="route(i.route, employee.id)"
-               class="mr-1 flex cursor-pointer items-center justify-center whitespace-nowrap rounded-md border px-4 py-2 text-sm font-medium transition"
-                :class="{
-                  'border-rose-500 bg-rose-500 text-white shadow-sm hover:bg-rose-600 dark:border-rose-500 dark:bg-rose-600' : route().current() === i.route,
-                  'border-slate-300 bg-white text-slate-600 hover:border-rose-300 hover:bg-rose-50 hover:text-rose-700 dark:border-slate-600 dark:bg-slate-700 dark:text-slate-200 dark:hover:border-rose-500 dark:hover:bg-slate-600' : route().current() !== i.route
-                }"
-          >
-            <span v-text="i.label"></span>
-          </Link>
-        </template>
-      </div>
-
-      <!--Select Navigation-->
-      <div class="block md:hidden">
-        <SelectInput v-model="selectedTab" @changed="changePage" :options="tabs" option-key="route"></SelectInput>
-      </div>
+      <DetailPageTabs :tabs="tabs" :route-param="employee.id"/>
       <!--Content-->
       <div class="my-6">
         <slot></slot>
