+++ resources/js/Pages/Modules/MeasurementDevice/Device/ShowPage.vue
@@ -1,11 +1,11 @@
 <script setup>
 import AppLayout from "@/Layouts/AppLayout.vue";
-import {Link, router, useForm} from "@inertiajs/vue3";
+import {router, useForm} from "@inertiajs/vue3";
 
 // Components
 import SimpleButton from "@/Components/Button/SimpleButton.vue"
-import SelectInput from "@/Components/Form/SelectInput.vue";
 import HelpButton from "@/Components/Help/HelpButton.vue";
+import DetailPageTabs from "@/Components/Navigation/DetailPageTabs.vue";
 import Modal from "@/Components/Modal/Modal.vue";
 import Form from "@/Components/Form/Form.vue";
 import FormSection from "@/Components/Form/FormSection.vue";
@@ -54,11 +54,6 @@ const submitDecommission = async () => {
 
 const reactivate = () => {
   router.post(route('measurement-device.reactivate', props.measurementDevice.id), {}, {preserveScroll: true})
-}
-
-const selectedTab = ref(route().current());
-const changePage = ()=>{
-  router.visit(route(selectedTab.value, props.measurementDevice.id));
 }
 
 /*Tabs*/
@@ -138,25 +133,7 @@ const handleDelete = () => {
     </template>
 
     <div>
-      <!--Tabs-->
-      <div class="hidden md:flex select-none pb-3 -mb-3 scrollbar-thin scrollbar-thumb-rose-500">
-        <template v-for="i in tabs">
-          <Link :href="route(i.route, measurementDevice.id)"
-               class="flex justify-center items-center text-sm font-medium mr-0.5 first:rounded-l-md last:rounded-r-md  px-4 py-2 whitespace-nowrap transition cursor-pointer"
-                :class="{
-                  'bg-rose-500/75 hover:bg-rose-800' : route().current() === i.route,
-                  'bg-slate-700 hover:bg-rose-500/50' : route().current() !== i.route
-                }"
-          >
-            <span v-text="i.label"></span>
-          </Link>
-        </template>
-      </div>
-
-      <!--Select Navigation-->
-      <div class="block md:hidden">
-        <SelectInput v-model="selectedTab" @changed="changePage" :options="tabs" option-key="route"></SelectInput>
-      </div>
+      <DetailPageTabs :tabs="tabs" :route-param="measurementDevice.id"/>
       <!--Content-->
       <div class="my-6">
         <slot></slot>
